# api/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
    SessionLoginRequest:
      additionalProperties: false
      properties:
        api_key:
          type: string
      required:
        - api_key
      type: object
    SessionStatus:
      additionalProperties: true
      properties:
        auth_mode:
          enum:
            - loopback
            - api_key
            - session
            - required
          type: string
        csrf_token:
          type: string
        google_oidc_enabled:
          type: boolean
        https:
          type: boolean
        plain_http_warning:
          type: boolean
      required:
        - auth_mode
        - https
        - plain_http_warning
        - google_oidc_enabled
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.36.0
openapi: 3.1.0
paths:
...
  /api/session:
    delete:
      operationId: logoutSession
      responses:
        "204":
          description: No Content
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      summary: Delete the current browser session
      tags:
        - Session
    get:
      operationId: getSession
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      summary: Get browser authentication status
      tags:
        - Session
  /api/session/login:
    post:
      operationId: loginSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionLoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      summary: Create an in-memory browser session
      tags:
        - Session
...